Modularize express app and make server easier to read
