+++ server/app/config/app.js
@@ -1,0 +1,24 @@
+import express from 'express';
+import bodyParser from 'body-parser';
+import morgan from 'morgan';
+
+import routes from './../routes/';
+
+const app = express();
+const router = express.Router();
+
+app.use(morgan('combined'));
+
+app.use(bodyParser.urlencoded({ extended: true }));
+app.use(bodyParser.json());
+
+routes(router);
+
+// Fail safe for routes that don't exist
+router.get('*', (req, res) => {
+    res.send({ message: 'You may be lost! Go home'})
+})
+
+app.use('/api', router);
+
+export default app;
